+++ Utilities/Exceptions/ItemNotFoundException.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace Utility.Exception
+namespace Utilities.Exceptions
 {
     /// <summary>
     /// Signs that sequence doesn't contain any items that one was looking for.
